[forgejo] Example patch for fixing ArgoCD webhooks
ArgoCD can be configured to receive webhooks, but since the ROOT_URL
doesn't match the URL used for Argo to reach Forgejo, the webhooks don't
get trigger applications that use the repo (because the URLs don't
match).  This just shows how to update the ROOT_URL to match what Argo
expects.  You still need to configure the webhook on any repo used.