[adguard] Add backup definition
Similar to the forgejo one, this just copies the state files off the PVC
and into a tarball.  Right now it's everything, but probably worth
culling in case there's logs or such that will grow and hold no recovery
value.

Also update the secret request to hint at the expected address value