[forgejo] Added note around weird hack
In order to incorporate local stuff that argo references (which uses
internal DNS resolution for the service), added an example for how to
temporarily define the same DNS record exterally (i.e. via ingress)
without leaking.  This is meant to be manually maintained (as
documented) but wanted to document it.