[redis] Add redis-controller
This is a controller+CRD for defining redis nodes.  There is a lot of
exploration to do here as I don't need a lot of what is offered (like
persistence) right now.  Tried to use the kustomize target for the
upstream but it's very out of date, so instead took the helm output and
cleaned up for use.

This is mostly needed for some services that like to offload work to
worker nodes and use redis for the brokering.