[immich] Initial definition
This is in the initial adoption testing... a photo manager option that
expects for a PVC to be used for the actual storage.  The PVC
expectations are around having external data resiliency provided on that
provider (like say a NAS with backups for a drive).  This will probably
get tweaked as I try adopting properly beyond just PoC deployments.